project has been broken down into modules
